more plumbing for client tests
