+++ client/src/registry.ts
@@ -1,24 +1,24 @@
-// TODO:
-// isHuman(Address)
-// signIdentity(IdentityDefinition)
-// getHuman(Address): IdentityDefinition
-// getHumans(): Observable<IdentityDefinition[]>
-
 import {
-  getEnabledWeb3
-} from "./web3";
+  getEnabledWeb3,
+  getNetworkName
+} from "./web3Utils";
 import {
   sendTransaction,
   toIOperationObservable,
   Web3Receipt
 } from "./transactionUtils";
 
+const abi = require("@dorgtech/id-dao-contracts/build/contracts/IdentityRegistry.json");
+const addresses = require("@dorgtech/id-dao-contracts/migrations/registries.json");
+
 let registry: any = undefined;
 
 export const getRegistry = async (): Promise<any> => {
   if (!registry) {
     const web3 = await getEnabledWeb3();
-    registry = new web3.eth.Contract(abi, address);
+    registry = new web3.eth.Contract(
+      abi, addresses[await getNetworkName()]
+    );
   }
 
   return registry;
@@ -39,3 +39,9 @@ export const removeSelf = async (): Promise<boolean | Error> => {
   const observable = sendTransaction(registry.methods.removeSelf(), map);
   return await toIOperationObservable(observable).send();
 }
+
+// TODO:
+// isHuman(Address)
+// signIdentity(IdentityDefinition)
+// getHuman(Address): IdentityDefinition
+// getHumans(): Observable<IdentityDefinition[]>
