+++ client/test/registry.test.ts
@@ -1,0 +1,2 @@
+// TODO:
+// removeSelf
